* develop:
  fix: flake8
  fix: TypeError: 'float' object cannot be interpreted as an integer
  fix: 'Settings' object attribute 'set' is read-only
  fix: assertEquals was removed in 3.12

+++ nv/commands.py
@@ -3068,7 +3068,7 @@ class nv_vi_move_to_file_percent(TextCommand):
     def run(self, edit, mode=None, count=None):
         if not count: # Go to {count} percentage in the file, on the first non-blank in the line linewise. To compute the new line number this formula is used: ({count} * number-of-lines + 99) / 100
             return
-        row = self.view.rowcol(self.view.size())[0] * (count / 100)
+        row = int(self.view.rowcol(self.view.size())[0] * (count / 100))
         def f(view, s):
             return Region(view.text_point(row, 0))
         regions_transformer(self.view, f)
